Coloquei a TabBar no Cabeçalho
